refactor: consolidate packet list bounds onto PacketGuard, drop jukebox rate limit
Move the ergonomic isValidIntList helper into the canonical com.eu.habbo.util.PacketGuard
and delete the duplicate com.eu.habbo.messages.incoming.PacketListGuard so there is one
packet-bounds utility. The jukebox, secret-craft, and available-recipes handlers now call
PacketGuard; the unit coverage moves to PacketGuardTest.

Remove the 250ms rate limit override on JukeBoxRequestTrackDataEvent: the count/remaining-byte
guard already stops the allocation DoS, and the silent drop risked blanking track names for
legitimate clients that request track data twice in quick succession.
